+++ src/routes/dashboard.rs
@@ -59,14 +59,15 @@ pub async fn view_dashboard(
                                 )
                                 .await,
                         );
-                        result.push_str(format!("<a href=\"/delete?msgid={}\">Delete</a>\n", msgid).as_str());
+                        result.push_str(
+                            format!("<a href=\"/delete?msgid={}\">Delete</a>\n", msgid).as_str(),
+                        );
                         result.push('\n');
                     }
                 }
                 result
             }
             _ => {
-                // anything other than 'msg' is treated as a submessage
                 let msg_vec = client.get_flagged_submessages().await;
                 let mut result = "".to_string();
                 if let Ok(v) = msg_vec {
@@ -74,7 +75,13 @@ pub async fn view_dashboard(
                         let parentid = i.parent_msg.clone();
                         let submsgid = i.submsg_id.clone();
                         result.push_str(&data.formatter.format_into_submessage(i).await);
-                        result.push_str(format!("<a href=\"/delete?msgid={}&submsgid={}\">Delete</a>\n", parentid, submsgid).as_str());
+                        result.push_str(
+                            format!(
+                                "<a href=\"/delete?msgid={}&submsgid={}\">Delete</a>\n",
+                                parentid, submsgid
+                            )
+                            .as_str(),
+                        );
                         result.push('\n');
                     }
                 }
@@ -98,7 +105,9 @@ pub async fn login_page(
     session: actix_session::Session,
     form: web::Form<LoginForm>,
 ) -> impl Responder {
-    if form.password == data.config.admin_password {
+    // the password is stored as a hash
+    let hashed = sha256::digest(&form.password);
+    if hashed == data.config.admin_password {
         let session_insert_result = session.insert("logged_in", true);
         match session_insert_result {
             Ok(_) => log::info!("Admin successfully authorized"),
@@ -121,6 +130,36 @@ pub async fn delete_msg(
         if !logged_in {
             return web::Redirect::to("/error?error_code=403").see_other();
         }
+    } else {
+        return web::Redirect::to("/error?error_code=403").see_other();
     }
-    return web::Redirect::to("/b").see_other();
+    let client = data.db_client.lock().await;
+    if let Some(submsgid) = query.submsgid {
+        let row = client.get_single_submessage(query.msgid, submsgid).await.unwrap();
+        let image_paths = row.image.split(';').collect::<Vec<&str>>();
+        purge_images(image_paths);
+        client.delete_submsg(query.msgid, submsgid).await;
+    } else {
+        let row = client.get_single_message(query.msgid).await.unwrap();
+        let submsg_vec = client.get_submessages(query.msgid).await;
+        if let Ok(v) = submsg_vec {
+            for i in v {
+                let image_paths_sub = i.image.split(';').collect::<Vec<&str>>();
+                purge_images(image_paths_sub);
+            }
+        }
+        let image_paths = row.image.split(';').collect::<Vec<&str>>();
+        purge_images(image_paths);
+        client.delete_msg(query.msgid).await;
+    }
+    return web::Redirect::to("/dashboard").see_other();
+}
+
+fn purge_images(paths: Vec<&str>) {
+    for i in paths {
+        match std::fs::remove_file(std::path::Path::new(&format!("{}/{}", env!("CARGO_MANIFEST_DIR"), i))) {
+            Ok(_) => log::debug!("Deleted media file: {}", i),
+            Err(_) => log::error!("Media file deletion failed: {}", i),
+        };
+    }
 }